First off, and most important, is that I changed the format of how the file is read. The new format is 	X values 	number of repeated measurements 	A values 	Y values (continuing to end of file)
I did this to ensure that A values are not
consumed by over parsing the Y values. I
also did it to generate arrays with all
the sizes pre-determined, such as in the
case where there are three Y measurements
for A=2. The end result is that Y[i] will
correspond to A[i]. Across all of these
it is assumed that X will not change,
hence why X[i] is not passed to minimize()

Secondly, I haven't managed to fix the error
"ValueError: operands could not be broadcast together with shapes (50,) (3,) "

This occurs whenever A has more than one value,
regardless of passing a single value (ex A[0])
to minimize(), as shown by the second number,
3 in this case, changing to match whatever A
is. X will always be constant when this error
occurs, at 50.

For some reason, emfret() will be called with
the correct values of X and A up until the last
iteration, at which point X jumps to 50, and A
jumps to be whatever number of elements it has.

Tracking the values of X and Y in an older
version of fret.py will show that the same
thing occurs, however, because sizeof(A)=1
the result is 50x1. I don't fully understand
why this is.

This error will prevent generation of the
graph because it occurs before that point,
however, the program does appear to run
correctly beyond that.